requirements.txt
Jinja2==3.0.3
MarkupSafe==2.0.1
PyYAML==6.0
